fix: remove newlines in link descriptions
Currently, when an orgmode link has a newline in its description, it
is passed to the description of the Gemini link. Example:

```org
[[https://example.com][This is a link with a
newline to example.com]]
```

Which exports to:

```gmi
=> https://example.com This is a link with a
newline to example.com
```

This commit fixes this by replacing newlines with a simple space. The
new output is:

```gmi
=> https://example.com This is a link with a newline to example.com
```

It also fixes a compilation error on Emacs 30 by properly escaping
quotation marks.

Signed-off-by: Lucien Cartier-Tilet <lucien@phundrak.com>